+++ src/database/sql/create_map.sql
@@ -1,8 +1,11 @@
+-- DROP TABLE IF EXISTS map
 CREATE TABLE IF NOT EXISTS map (
-    pid        INTEGER NOT NULL,
-    addr_begin INTEGER NOT NULL,
-    addr_end   INTEGER NOT NULL,
+    pid        BIGINT NOT NULL,
+    addr_begin BIGINT NOT NULL,
+    addr_end   BIGINT NOT NULL,
     perm       TEXT    NOT NULL,
     path       TEXT    NOT NULL
 );
-CREATE INDEX IF NOT EXISTS map_pid_index ON map(pid); 
+
+-- DROP INDEX IF EXISTS map_pid_index
+CREATE INDEX IF NOT EXISTS map_pid_index ON map(pid);
